refactor code v1.5.5
