+++ src/rust_analyzer_config.ts
@@ -4,7 +4,7 @@ import { isValidToolchain, removeChannel, updateChannel } from './toolchain';
 import { log } from './logger';
 import { findCargoToml } from './cargo_manifest';
 
-export const updateRustAnalyzerConfig = vscode.commands.registerCommand('cargo.rust-analyzer.config', async () => {
+export const rustAnalyzerConfig = vscode.commands.registerCommand('cargo.rust-analyzer.config', async () => {
     const args = await vscode.window.showInputBox({
         prompt: 'Add or Remove Rust Analyzer Config'
     });
