+++ src/symbol_kind.ts
